Cast water_level where it meets a float
Three kernels multiply pm->water_level, a pm_water_level_t, into a float
expression to scale water friction. Clang's -Wenum-float-conversion, which
the Xcode build enables and the autotools build does not, warns on each.

The promotion was already happening and 0 through 3 are exact in a float,
so this is the cast the compiler was going to apply, written down.

+++ src/game/common/bg_pmove_quake.c
@@ -317,7 +317,7 @@ static void Pm_QuakeFriction(void) {
   float drop = 0.f;
 
   if (pm->water_level >= WATER_WAIST) {
-    drop = speed * pm->s.params.friction_water * pm->water_level * pm_locals.time;
+    drop = speed * pm->s.params.friction_water * (float) pm->water_level * pm_locals.time;
   } else if (pm->s.flags & PMF_ON_GROUND) {
     const float control = Maxf(speed, pm->s.params.speed_stop);
     drop = control * friction * pm_locals.time;
